requirements.txt
argparse==1.2.1
dateutils==0.6.6
distribute==0.6.24
ipython==0.13.2
logilab-astng==0.24.3
logilab-common==0.59.1
numpy==1.7.1
pandas==0.11.0
pep8==1.4.6
pylint==0.28.0
pyredmine==0.2.4
python-dateutil==2.1
pytz==2013b
six==1.3.0
wsgiref==0.1.2
